yaml file work
I have the base line implementation for being able to open yaml files within a certain directory and to display their contents! I have two yaml files that I tested this with. They are inside the "Playbooks" directory.

I like what I have so far. The implementation is as follows:

eagle add_user.yml

*displays contents*

eagle remove_user.yml

*displays contents*

you have to put ".yml" at the end of the respective file name and the files have to be in the "Playbook" directory that I made. Use the keyword "eagle" followed by the name of the yaml file that you are trying to run.