Successfully implemented baseline for inventory
I have successfully completed a baseline implementation for inventory management. All I have done is taken the IP's out of the playbook and put them in a separate file, parsed them, and passed them through the playbook execution. There is still more to be done and much to clean up.